requirements.txt
tensorflow==1.15.0
Pillow==7.0.0
numpy